renamed playerScoreApi to scoreApi
